# Conflicts:
#	com.unity.render-pipelines.universal/Editor/ShaderGUI/BaseShaderGUI.cs
#	com.unity.render-pipelines.universal/Editor/ShaderGraph/Targets/UniversalTarget.cs
#	com.unity.render-pipelines.universal/Editor/ShaderGraph/UniversalFields.cs

+++ com.unity.render-pipelines.universal/Editor/ShaderGUI/BaseShaderGUI.cs
@@ -215,7 +215,8 @@ namespace UnityEditor
             EditorGUI.BeginChangeCheck();
 
             m_SurfaceOptionsFoldout.value = EditorGUILayout.BeginFoldoutHeaderGroup(m_SurfaceOptionsFoldout.value, Styles.SurfaceOptions);
-            if(m_SurfaceOptionsFoldout.value){
+            if (m_SurfaceOptionsFoldout.value)
+            {
                 DrawSurfaceOptions(material);
                 EditorGUILayout.Space();
             }
@@ -335,7 +336,7 @@ namespace UnityEditor
             }
         }
 
-        public virtual void DrawAdditionalFoldouts(Material material){}
+        public virtual void DrawAdditionalFoldouts(Material material) {}
 
         public virtual void DrawBaseProperties(Material material)
         {
@@ -435,7 +436,7 @@ namespace UnityEditor
             SetupMaterialBlendMode(material);
 
             // Receive Shadows
-            if(material.HasProperty("_ReceiveShadows"))
+            if (material.HasProperty("_ReceiveShadows"))
                 CoreUtils.SetKeyword(material, "_RECEIVE_SHADOWS_OFF", material.GetFloat("_ReceiveShadows") == 0.0f);
 
             // Emission
@@ -462,7 +463,7 @@ namespace UnityEditor
                 throw new ArgumentNullException("material");
 
             bool alphaClip = false;
-            if(material.HasProperty("_AlphaClip"))
+            if (material.HasProperty("_AlphaClip"))
                 alphaClip = material.GetFloat("_AlphaClip") >= 0.5;
 
             if (alphaClip)
@@ -476,24 +477,24 @@ namespace UnityEditor
 
             if (material.HasProperty("_Surface"))
             {
-                SurfaceType surfaceType = (SurfaceType) material.GetFloat("_Surface");
+                SurfaceType surfaceType = (SurfaceType)material.GetFloat("_Surface");
                 if (surfaceType == SurfaceType.Opaque)
                 {
                     if (alphaClip)
                     {
-                        material.renderQueue = (int) RenderQueue.AlphaTest;
+                        material.renderQueue = (int)RenderQueue.AlphaTest;
                         material.SetOverrideTag("RenderType", "TransparentCutout");
                     }
                     else
                     {
-                        material.renderQueue = (int) RenderQueue.Geometry;
+                        material.renderQueue = (int)RenderQueue.Geometry;
                         material.SetOverrideTag("RenderType", "Opaque");
                     }
 
-                    material.renderQueue += material.HasProperty("_QueueOffset") ? (int) material.GetFloat("_QueueOffset") : 0;
-                    material.SetInt("_SrcBlend", (int) UnityEngine.Rendering.BlendMode.One);
-                    material.SetInt("_DstBlend", (int) UnityEngine.Rendering.BlendMode.Zero);
-                    material.SetInt("_ZWrite", 1);
+                    material.renderQueue += material.HasProperty("_QueueOffset") ? (int)material.GetFloat("_QueueOffset") : 0;
+                    material.SetFloat("_SrcBlend", (float)UnityEngine.Rendering.BlendMode.One);
+                    material.SetFloat("_DstBlend", (float)UnityEngine.Rendering.BlendMode.Zero);
+                    material.SetFloat("_ZWrite", 1.0f);
                     material.DisableKeyword("_ALPHAPREMULTIPLY_ON");
                     material.DisableKeyword("_ALPHAMODULATE_ON");
                     material.DisableKeyword("_PRESERVE_SPECULAR");
@@ -504,7 +505,7 @@ namespace UnityEditor
                 }
                 else
                 {
-                    BlendMode blendMode = (BlendMode) material.GetFloat("_Blend");
+                    BlendMode blendMode = (BlendMode)material.GetFloat("_Blend");
 
                     #if BLEND_DO_BRANCH
                     if (blendMode == (BlendMode) s_BlendModeOpaque)
@@ -590,9 +591,9 @@ namespace UnityEditor
 
                     // General Transparent Material Settings
                     material.SetOverrideTag("RenderType", "Transparent");
-                    material.SetInt("_ZWrite", 0);
+                    material.SetFloat("_ZWrite", 0.0f);
                     material.renderQueue = (int)RenderQueue.Transparent;
-                    material.renderQueue += material.HasProperty("_QueueOffset") ? (int) material.GetFloat("_QueueOffset") : 0;
+                    material.renderQueue += material.HasProperty("_QueueOffset") ? (int)material.GetFloat("_QueueOffset") : 0;
                     material.SetShaderPassEnabled("ShadowCaster", false);
                 }
             }
@@ -707,7 +708,7 @@ namespace UnityEditor
                     return properties[index];
             }
             if (propertyIsMandatory)
-                throw new ArgumentException("Could not find MaterialProperty: '" + propertyName + "', Num properties: " + (object) properties.Length);
+                throw new ArgumentException("Could not find MaterialProperty: '" + propertyName + "', Num properties: " + (object)properties.Length);
             return null;
         }
